refactor: Reorganize classes into package
Application will now only have 3 root packages:
- app: Everything that concerns the whole app (app navigation, app
entry points, root composable...)
- feat: Application's features, with each immediate sub-package
encapsulating to a feature group
- core: Classes and composables that are used across different features
 (repositories,models, services, utils, UI components...)

Reference: https://github.com/android/nowinandroid/blob/main/docs/ModularizationLearningJourney.md
