+++ app/src/main/java/com/haidoan/android/stren/app/ui/StrenApp.kt
@@ -1,19 +1,20 @@
 package com.haidoan.android.stren.app
 
-import androidx.compose.animation.ExperimentalAnimationApi
+import android.util.Log
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.*
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.haidoan.android.stren.designsystem.component.BottomNavigationBar
+import com.haidoan.android.stren.app.ui.StrenAppState
+import com.haidoan.android.stren.app.ui.rememberStrenAppState
+import com.haidoan.android.stren.core.designsystem.component.BottomNavigationBar
 
 private const val TAG = "StrenApp"
 val LocalSnackbarHostState =
     compositionLocalOf<SnackbarHostState> { error("No SnackbarHostState provided") }
 
-@OptIn(ExperimentalAnimationApi::class)
 @Composable
 fun StrenApp(
     viewModel: StrenAppViewModel = hiltViewModel(),
@@ -22,6 +23,7 @@ fun StrenApp(
     val snackbarHostState = remember { SnackbarHostState() }
     val isUserSignedIn by rememberSaveable { viewModel.isUserSignedIn }
 
+    Log.d(TAG, "isUserSignedIn: $isUserSignedIn")
     // This allows any screen in the composition to access snackbar
     CompositionLocalProvider(
         LocalSnackbarHostState provides snackbarHostState
